bookbridge/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
sqlalchemy
alembic
bcrypt
itsdangerous
httpx
python-multipart
groq
